+++ stroom-dashboard-client/src/main/java/stroom/dashboard/client/vis/MessageSupport.java
@@ -76,10 +76,9 @@ public class MessageSupport implements FrameListener, HasHandlers {
         final String functionName = JSONUtil.getString(message.get("functionName"));
 
         if ("link".equals(functionName)) {
-            final String title = JSONUtil.getString(message.get("title"));
             final String href = JSONUtil.getString(message.get("href"));
             final String target = JSONUtil.getString(message.get("target"));
-            final Hyperlink hyperlink = new Builder().title(title).href(href).type(target).build();
+            final Hyperlink hyperlink = new Builder().href(href).type(target).build();
             HyperlinkEvent.fire(this, hyperlink);
 
         } else {