+++ CHANGELOG.md
@@ -6,9 +6,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-* Issue **#986**: Fix direct dasbhoard links
+## [v6.0-beta.21] - 2018-12-13
 
-* Issue **#1006** : Added Exception Mapper for PermissionExceptions to return HTTP FORBIDDEN
+* Issue **#1022** : Added `Contains` condition to all search expression fields so that regex terms can be used.
+
+* Issue **#1024** : Superseded output helper no longer expects initialisation in all cases.
+
+## [v6.0-beta.20] - 2018-12-13
+
+* Issue **#1021** : Multiple changes to improve vis, dashboard and external linking in Stroom.
+
+* Issue **#1019** : Fix visualisations settings dialog so you can un-set text and list controls.
+
+* Issue **#986** : Fix direct dashboard links.
+
+* Issue **#1006** : Added Exception Mapper for PermissionExceptions to return HTTP FORBIDDEN.
 
 ## [v6.0-beta.19] - 2018-12-11
 
@@ -1367,7 +1379,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Issue **#202** : Initial release of the new data retention policy functionality.
 
-[Unreleased]: https://github.com/gchq/stroom/compare/v6.0-beta.19...6.0
+[Unreleased]: https://github.com/gchq/stroom/compare/v6.0-beta.21...6.0
+[v6.0-beta.21]: https://github.com/gchq/stroom/compare/v6.0-beta.20...v6.0-beta.21
+[v6.0-beta.20]: https://github.com/gchq/stroom/compare/v6.0-beta.19...v6.0-beta.20
 [v6.0-beta.19]: https://github.com/gchq/stroom/compare/v6.0-beta.18...v6.0-beta.19
 [v6.0-beta.18]: https://github.com/gchq/stroom/compare/v6.0-beta.17...v6.0-beta.18
 [v6.0-beta.17]: https://github.com/gchq/stroom/compare/v6.0-beta.16...v6.0-beta.17